Refractory  (#7)
* исправил надписи в таблицах описывающих работу алгоритма

* Рефакторинг и комментирование кода

* Добавил перечисляемые серии.

* repath for series merge

* Сделал Серию дженериками.
теперь она возвращает объект TapeReturn. Нужно пересмотреть логику объектов ***Return<T>

* fix crash when file is exists

* merge two **Return into UtilType.

+++ src/Algorithm/Sort/External/SeriesMerge/ExternalSeriesMergeSort.cs
@@ -1,7 +1,8 @@
-﻿using algorithms_cs.Serial;
-using algorithms_cs.Tape;
+﻿using algorithms_cs.Utils;
+using algorithms_cs.Utils.Serial;
+using algorithms_cs.Utils.Tape;
 
-namespace algorithms_cs.Algorithm.Sort.External.Merge;
+namespace algorithms_cs.Algorithm.Sort.External.SeriesMerge;
 
 public class MultiwaySort: Sort
 {
@@ -9,7 +10,7 @@ public class MultiwaySort: Sort
     private readonly string _tempDirectory;
     private readonly int _n;
     private RotatingDominoes _dominoes;
-    private List<Tape.Tape> _tapes;
+    private List<Tape> _tapes;
     private readonly string _templateNameFiles;
     
     public MultiwaySort(int N, string sourceFilePath)
@@ -25,7 +26,7 @@ public class MultiwaySort: Sort
         _templateNameFiles = _tempDirectory + "tempfile-{0}.multiwaymergesort";
         Directory.CreateDirectory(_tempDirectory);
         
-        _tapes = new List<Tape.Tape>();
+        _tapes = new List<Tape>();
         _n = N;
         
         var firstFilenames = new List<string>();
@@ -57,7 +58,7 @@ public class MultiwaySort: Sort
 
     private void TrashDelete(string outFilename)
     {
-        File.Copy(outFilename, _sourceFilePath+"-sorted");
+        if (!File.Exists(_sourceFilePath+"-sorted")) File.Copy(outFilename, _sourceFilePath+"-sorted");
         foreach (var filename in _dominoes.ReadFilenames)
         {
             File.Delete(filename);
@@ -76,26 +77,25 @@ public class MultiwaySort: Sort
         var tapeWrites = _dominoes.WriteFilenames.Select(filename => new TapeWriter<double>(filename)).ToList();
 
         var indexTape = 0;
-        if (indexTape >= _n) throw new IndexOutOfRangeException("indexTape out of range tapeWrites");
         
         var s1 = new Series(initTape);
-        SeriesReturn<double> value;
+        UtilReturn<double> value;
 
         do
         {
             do
             {
                 value = s1.Next();
-                if (value.GetType() == SeriesReturnType.Correct)
+                if (value.GetType() == UtilReturnType.Correct)
                 {
                     tapeWrites[indexTape].Write(value.GetValue());
                 }
-            } while (value.GetType() == SeriesReturnType.Correct);
+            } while (value.GetType() == UtilReturnType.Correct);
 
             indexTape += 1;
             if (indexTape >= _n) indexTape = 0;
             s1 = new Series(initTape);
-        } while (value?.GetType() != SeriesReturnType.TapeEnded);
+        } while (value?.GetType() != UtilReturnType.TapeEnded);
         
         foreach (var tape in tapeWrites)
         {
@@ -131,17 +131,17 @@ public class MultiwaySort: Sort
         var multipleSeries = new List<Series>();
         multipleSeries.AddRange(readTapes.Select(tape => new Series(tape)));
         var collector = new Collector(multipleSeries);
-        SeriesReturn<double> value;
+        UtilReturn<double> value;
         do
         {
             do
             {
                 value = collector.Next();
-                if (value.GetType() == SeriesReturnType.Correct)
+                if (value.GetType() == UtilReturnType.Correct)
                 {
                     writeTapes[indexTape].Write(value.GetValue());
                 }
-            } while (value.GetType() == SeriesReturnType.Correct);
+            } while (value.GetType() == UtilReturnType.Correct);
 
             count++;
             if (count > 2) count = 2;
